factor out router
